Generate code without implicit casts
Gather up the enum types and pass them to the `fromParams` methods so
that an appropriate explicit cast can be inserted - either to the
primitive type for an enum, or to `'Map'` for message types.

Refactor to collection literals.

Drop an unnecessary `new`.

Update goldens.

+++ test/goldens/enum.dart
@@ -9,7 +9,7 @@ class MessageUsingEnum {
 
   factory MessageUsingEnum.fromJson(Map params) => MessageUsingEnum._(
       params.containsKey('enumField') && params['enumField'] != null
-          ? SomeEnum.fromJson(params['enumField'])
+          ? SomeEnum.fromJson((params['enumField'] as int))
           : null);
 
   final SomeEnum enumField;
